AbstractProcessorModel eingeführt
